Update README.md
Hey, I am doing a project for a class I am taking. We are learning about github and how to contribute. I love Unity and decided to update your project README to include a "contribution standards" and "Code of Conduct". If you reject the update that is fine, it will not hurt my feelings. I just need to prove that I attempted to submit a change to documentation.

Thanks, 
Mike